+++ angular-workspace/projects/trimble-oss/modus-angular-components/src/lib/stencil-generated/components.ts
@@ -247,6 +247,42 @@ export class ModusChip {
 }
 
 
+export declare interface ModusContentTreeItem extends Components.ModusContentTreeItem {
+  /**
+   * An event that first on item checkbox click 
+   */
+  checkboxClick: EventEmitter<CustomEvent<boolean>>;
+  /**
+   * An event that fires on item expand click 
+   */
+  expandClick: EventEmitter<CustomEvent<boolean>>;
+  /**
+   * An event that fires on item click 
+   */
+  itemClick: EventEmitter<CustomEvent<any>>;
+
+}
+
+@ProxyCmp({
+  defineCustomElementFn: undefined,
+  inputs: ['allowDrag', 'checkboxChecked', 'depth', 'disabled', 'expandable', 'expanded', 'includeBottomBorder', 'includeCheckbox', 'includeFolderIcon', 'selected', 'size']
+})
+@Component({
+  selector: 'modus-content-tree-item',
+  changeDetection: ChangeDetectionStrategy.OnPush,
+  template: '<ng-content></ng-content>',
+  inputs: ['allowDrag', 'checkboxChecked', 'depth', 'disabled', 'expandable', 'expanded', 'includeBottomBorder', 'includeCheckbox', 'includeFolderIcon', 'selected', 'size']
+})
+export class ModusContentTreeItem {
+  protected el: HTMLElement;
+  constructor(c: ChangeDetectorRef, r: ElementRef, protected z: NgZone) {
+    c.detach();
+    this.el = r.nativeElement;
+    proxyOutputs(this, this.el, ['checkboxClick', 'expandClick', 'itemClick']);
+  }
+}
+
+
 export declare interface ModusDropdown extends Components.ModusDropdown {
   /**
    * An event that fires on dropdown close.