+++ stencil-workspace/src/components/icons/icon-chevron-up-thick.tsx
@@ -8,7 +8,11 @@ interface IconProps {
 }
 
 export const IconChevronUpThick: FunctionalComponent<IconProps> = (props: IconProps) => (
-  <svg class="icon-chevron-up-thick" height={props.size ?? 16} width={props.size ?? 16} onClick={props.onClick ? () => props.onClick() : null} viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
-    <path d="M12.5292 8.21649C12.2405 7.92783 11.7595 7.92783 11.4708 8.21649L5.21649 14.4387C4.92783 14.7594 4.92783 15.1014 5.21649 15.39L6 16C6.28866 16.2887 6.67927 16.2887 7 16L12 11L17 16C17.3207 16.2887 17.7572 16.5556 18.0458 16.2669L18.7835 15.5292C19.0722 15.2405 19.0722 14.7594 18.7835 14.4387L12.5292 8.21649Z" fill={props.color ?? '#6A6976'} />
+  <svg class="icon-chevron-up-thick" xmlns="http://www.w3.org/2000/svg" height={props.size ?? 16}
+       width={props.size ?? 16} onClick={props.onClick ? () => props.onClick() : null} fill="currentColor"
+       viewBox="0 0 32 32">
+    <path
+      d="M3.94 21.753A2.483 2.483 0 0 0 6 22.864c.523 0 1.024-.165 1.45-.476l.035-.026.032-.028L16 14.941l8.483 7.393.032.028.034.026c.427.311.928.476 1.45.476.824 0 1.594-.415 2.06-1.112.761-1.135.502-2.704-.573-3.519l-9.974-8.572-.03-.026-.032-.024a2.445 2.445 0 0 0-2.9 0l-.032.024-.03.026-9.974 8.572c-1.075.816-1.335 2.384-.574 3.52z"
+      fill={props.color ?? '#6A6976'}/>
   </svg>
 );